requirements.txt
zenml
transformers
datasets
peft
trl
bitsandbytes
accelerate
torch
faiss-cpu
sentence-transformers